+++ src/features/table/components/TableHeader/index.ts
@@ -2,5 +2,6 @@ export * from './TableHeader';
 export * from './TableHeaderRow';
 export * from './TableHeaderCell';
 export * from './FilterDropdown';
+export * from './FilterCalendar';
 export * from './TableHeaderSearchRow';
 export * from './TableHeaderSearchCell';